refactor(python): split profile and assignment owners
