+++ .buildchain/kfd/kfd-1/release-gate.json
@@ -106,13 +106,13 @@
       "standard": "KFD-1",
       "standardKey": "kfd-1",
       "result": "passed",
-      "preBuildWitnessSha256": "f5a3302b4c146e98019e9e163707ef48e790c8154f80253b933c7e4ff1c28f59",
+      "preBuildWitnessSha256": "c0f792391bb45793a567a43732581b67ac2ff483d029fc249929ee4b1bae74db",
       "sourceHashes": {
-        "sha256": "0e05deef475e826b324d4cac96b945ab5b6dd752300efccfc513fcfa6d3cd79e",
+        "sha256": "88731ea5bc94186c0265e1ec1159e247e3d3d36fa1a3d5d9fbeea9d047665863",
         "surfaceCount": 23
       },
       "artifactHashes": {
-        "sha256": "cb04a909938565aa4081d25cc349231b2c133bd9c3d8c592db3c4ebe874916a2",
+        "sha256": "9e2c151800374122b2d57505e773e484bcddbd7160f0d1f57c119e865c89fb99",
         "surfaceCount": 23
       },
       "witness": {
@@ -152,7 +152,7 @@
         "sourceVerificationRequired": true,
         "canonicalPolicy": {
           "path": "framework/contract/kungfu-agent-first-canonical-policy.json",
-          "sha256": "dfa0acd7118383e5f177602a8638314470a7a5ef65604e49baa4289e3f82c3f1"
+          "sha256": "ede207b8aa11fdc0c8bf0274a5cc29f14d265014905fe6a0a841736995f77032"
         },
         "registry": {
           "path": "framework/contract/kungfu-contracts.registry.json",
@@ -338,9 +338,9 @@
           {
             "name": "kungfu-primitive-catalog",
             "sourcePath": "framework/primitive/kungfu-primitive-catalog.contract.json",
-            "sourceSha256": "3bee9c11911295b142bbec8ef5fc0fc6cc98d45bd7675ea228dbeec7cfd07adc",
+            "sourceSha256": "edd8299e3162dc147d21f01ef6733c1beab0cc475f6bfffc2aaf3fe06cf74835",
             "artifactPath": "config/primitive/kungfu-primitive-catalog.contract.json",
-            "expectedSha256": "3bee9c11911295b142bbec8ef5fc0fc6cc98d45bd7675ea228dbeec7cfd07adc",
+            "expectedSha256": "edd8299e3162dc147d21f01ef6733c1beab0cc475f6bfffc2aaf3fe06cf74835",
             "byteForByte": true
           }
         ],
@@ -617,8 +617,8 @@
           {
             "name": "kungfu-primitive-catalog",
             "sourcePath": "framework/primitive/kungfu-primitive-catalog.contract.json",
-            "expectedSha256": "3bee9c11911295b142bbec8ef5fc0fc6cc98d45bd7675ea228dbeec7cfd07adc",
-            "actualSha256": "3bee9c11911295b142bbec8ef5fc0fc6cc98d45bd7675ea228dbeec7cfd07adc",
+            "expectedSha256": "edd8299e3162dc147d21f01ef6733c1beab0cc475f6bfffc2aaf3fe06cf74835",
+            "actualSha256": "edd8299e3162dc147d21f01ef6733c1beab0cc475f6bfffc2aaf3fe06cf74835",
             "status": "passed",
             "reason": ""
           }
@@ -873,10 +873,10 @@
           {
             "name": "kungfu-primitive-catalog",
             "sourcePath": "framework/primitive/kungfu-primitive-catalog.contract.json",
-            "sourceSha256": "3bee9c11911295b142bbec8ef5fc0fc6cc98d45bd7675ea228dbeec7cfd07adc",
+            "sourceSha256": "edd8299e3162dc147d21f01ef6733c1beab0cc475f6bfffc2aaf3fe06cf74835",
             "artifactPath": "config/primitive/kungfu-primitive-catalog.contract.json",
-            "expectedSha256": "3bee9c11911295b142bbec8ef5fc0fc6cc98d45bd7675ea228dbeec7cfd07adc",
-            "actualSha256": "3bee9c11911295b142bbec8ef5fc0fc6cc98d45bd7675ea228dbeec7cfd07adc",
+            "expectedSha256": "edd8299e3162dc147d21f01ef6733c1beab0cc475f6bfffc2aaf3fe06cf74835",
+            "actualSha256": "edd8299e3162dc147d21f01ef6733c1beab0cc475f6bfffc2aaf3fe06cf74835",
             "byteForByte": true,
             "status": "passed",
             "reason": ""
