+++ .buildchain/kfd/support-matrix.json
@@ -102,7 +102,7 @@
         "evidenceRoots": [
           {
             "path": ".buildchain/kfd/kfd-2/release-claims.json",
-            "sha256": "sha256:f0da873e4c48ef6561ebfb55798f71b4dd5dae15ea9ecd4499a1db30bf0a452d"
+            "sha256": "sha256:0bce8bb06409366402f0164cb17bde66ec1f10568ae6787b261e337c591881d2"
           },
           {
             "path": "docs/qualification/kfd2-trust-assessment.md",
@@ -274,7 +274,7 @@
           },
           {
             "path": "framework/primitive/kungfu-primitive-catalog.contract.json",
-            "sha256": "sha256:3bee9c11911295b142bbec8ef5fc0fc6cc98d45bd7675ea228dbeec7cfd07adc"
+            "sha256": "sha256:edd8299e3162dc147d21f01ef6733c1beab0cc475f6bfffc2aaf3fe06cf74835"
           }
         ]
       },
@@ -635,7 +635,7 @@
         "evidenceRoots": [
           {
             "path": "framework/core/src/python/kungfu/assignment_orchestration.py",
-            "sha256": "sha256:009cebcd0abb38435421181c91900ed615a8fbd4e273d2cde6b84cb4073b9b9f"
+            "sha256": "sha256:57a91ade36600a6b39951951216fa3a9c660dafc1ea54462c89f16f1da5a95fa"
           }
         ]
       },
