+++ config/primitive/kungfu-primitive-catalog.contract.json
@@ -1593,7 +1593,7 @@
               },
               {
                 "path": "framework/core/src/python/kungfu/assignment_orchestration.py",
-                "root": "sha256:009cebcd0abb38435421181c91900ed615a8fbd4e273d2cde6b84cb4073b9b9f"
+                "root": "sha256:57a91ade36600a6b39951951216fa3a9c660dafc1ea54462c89f16f1da5a95fa"
               }
             ],
             "policy": "all four language implementation references are required for complete"
@@ -1635,7 +1635,7 @@
               },
               {
                 "path": "framework/core/src/python/kungfu/assignment_orchestration.py",
-                "root": "sha256:009cebcd0abb38435421181c91900ed615a8fbd4e273d2cde6b84cb4073b9b9f"
+                "root": "sha256:57a91ade36600a6b39951951216fa3a9c660dafc1ea54462c89f16f1da5a95fa"
               }
             ],
             "policy": "implementation references are bounded operational evidence, not availability"
@@ -1669,7 +1669,7 @@
               },
               {
                 "path": "framework/core/src/python/kungfu/assignment_orchestration.py",
-                "root": "sha256:009cebcd0abb38435421181c91900ed615a8fbd4e273d2cde6b84cb4073b9b9f"
+                "root": "sha256:57a91ade36600a6b39951951216fa3a9c660dafc1ea54462c89f16f1da5a95fa"
               }
             ],
             "proofRefs": []
@@ -1721,5 +1721,5 @@
       "artifacts": []
     }
   ],
-  "catalogRoot": "sha256:e114a9293dfa9a61b07a5238c8f432de77bdd51216115774429015848140977a"
+  "catalogRoot": "sha256:795ed01db58ac4f1d809198d57ca173c68a751728ff6d53d85735dbe2a3c2f4b"
 }
