+++ developer/sdk/kfd/kfd-2/release-claims.json
@@ -11,7 +11,7 @@
     "version": "4.0.0-alpha.3",
     "channel": "repository-default",
     "tag": "v4.0.0-alpha.3",
-    "sourceSha": "adba36226f860c5f9fd29cd0a51ed9b09bdfeb92"
+    "sourceSha": "d6fcd79eb8dd6de51c58c883c756f2d9c3892eaf"
   },
   "claims": [
     {
