+++ framework/contract/kungfu-agent-first-canonical-policy.json
@@ -571,13 +571,13 @@
       },
       "source": {
         "path": "framework/primitive/kungfu-primitive-catalog.contract.json",
-        "sha256": "sha256:3bee9c11911295b142bbec8ef5fc0fc6cc98d45bd7675ea228dbeec7cfd07adc",
-        "renderedSha256": "sha256:3bee9c11911295b142bbec8ef5fc0fc6cc98d45bd7675ea228dbeec7cfd07adc",
+        "sha256": "sha256:edd8299e3162dc147d21f01ef6733c1beab0cc475f6bfffc2aaf3fe06cf74835",
+        "renderedSha256": "sha256:edd8299e3162dc147d21f01ef6733c1beab0cc475f6bfffc2aaf3fe06cf74835",
         "byteForByte": true
       },
       "artifact": {
         "path": "config/primitive/kungfu-primitive-catalog.contract.json",
-        "expectedSha256": "sha256:3bee9c11911295b142bbec8ef5fc0fc6cc98d45bd7675ea228dbeec7cfd07adc"
+        "expectedSha256": "sha256:edd8299e3162dc147d21f01ef6733c1beab0cc475f6bfffc2aaf3fe06cf74835"
       }
     }
   ]
