+++ framework/core/src/python/kungfu/_assignment_orchestration/__init__.py
@@ -1,0 +1,1 @@
+# SPDX-License-Identifier: Apache-2.0
