+++ framework/core/src/python/kungfu/cli/commands/assignment.py
@@ -6,16 +6,14 @@ from __future__ import annotations
 
 import json
 import os
-from datetime import UTC, datetime, timedelta
 from pathlib import Path
-import uuid
 
 import click
 
 from kungfu import assignment_orchestration as orchestration
 from kungfu import assignment_close
 from kungfu import assignment_evidence
-from kungfu import assignment_review_lifecycle
+from kungfu import assignment_review_lifecycle  # noqa: F401 -- private owner seam
 from kungfu import assignment_start
 from kungfu import initiative_family
 from kungfu.assignment_runtime import (
@@ -26,7 +24,7 @@ from kungfu.assignment_runtime import (
 from kungfu.assignment_runtime import profile_lifecycle
 from kungfu import dogfood as dogfood_api
 from kungfu import profile_sdk
-from kungfu.agent import run_agent
+from kungfu.agent import run_agent  # noqa: F401 -- private owner seam
 from kungfu.agent import resources as agent_resources
 from kungfu.cli.commands import (
     PrioritizedCommandGroup,
@@ -36,7 +34,9 @@ from kungfu.cli.commands import (
 from kungfu.cli.commands import assignment_review
 from kungfu.cli.commands import assignment_session
 from kungfu.cli.surface_contract import surface
-from kungfu.initiative_family import typed_v2 as initiative_family_v2
+from kungfu.initiative_family import (  # noqa: F401 -- private owner seam
+    typed_v2 as initiative_family_v2,
+)
 from kungfu.storage import service as storage_service
 from kungfu.workspace import prepare_workspace_write, resolve_workspace_target
 from kungfu.assignment_lifecycle.ports import AssignmentRuntime
@@ -540,622 +540,7 @@ assignment_session.register_finalize_agent_session_command(
 )
 
 
-@assignment.command(
-    name="start-plan",
-    help="preview admission, lease, kickoff, and one bound Agent run",
-)
-@click.argument(
-    "request_file", type=click.Path(exists=True, dir_okay=False, path_type=Path)
-)
-@click.option("--workspace", "workspace_root", type=click.Path(file_okay=False))
-@click.option("--home", is_flag=True)
-@click.option("--initiative-id", default="")
-@click.option("--assignment-id", default="")
-@click.option("--agent", "profile_id", required=True)
-@click.option("--actor", default="local-user", show_default=True)
-@click.option(
-    "--allow-foreign-binding",
-    is_flag=True,
-    help="development/testing override; retained as degraded provenance",
-)
-@assignment_context
-def start_plan(
-    ctx,
-    request_file,
-    workspace_root,
-    home,
-    initiative_id,
-    assignment_id,
-    profile_id,
-    actor,
-    allow_foreign_binding,
-):
-    _emit(
-        _run(
-            lambda: _work_start_plan(
-                config_home=ctx.config_home,
-                runtime_home=ctx.home,
-                request_file=request_file,
-                workspace_root=workspace_root,
-                home=home,
-                initiative_id=initiative_id,
-                assignment_id=assignment_id,
-                profile_id=profile_id,
-                actor=actor,
-                allow_foreign_binding=allow_foreign_binding,
-            )
-        )
-    )
-
-
-@assignment.command(
-    name="start",
-    help="admit, lease, kickoff, and launch one Agent from an exact plan",
-)
-@click.argument(
-    "request_file", type=click.Path(exists=True, dir_okay=False, path_type=Path)
-)
-@click.option("--workspace", "workspace_root", type=click.Path(file_okay=False))
-@click.option("--home", is_flag=True)
-@click.option("--initiative-id", default="")
-@click.option("--assignment-id", default="")
-@click.option("--agent", "profile_id", required=True)
-@click.option("--actor", default="local-user", show_default=True)
-@click.option("--expected-plan-root", required=True)
-@click.option("--execute", is_flag=True)
-@click.option("--events-json", is_flag=True)
-@click.option(
-    "--allow-foreign-binding",
-    is_flag=True,
-    help="development/testing override; retained as degraded provenance",
-)
-@assignment_context
-@surface(id="kungfu.work.start")
-def start_work(
-    ctx,
-    request_file,
-    workspace_root,
-    home,
-    initiative_id,
-    assignment_id,
-    profile_id,
-    actor,
-    expected_plan_root,
-    execute,
-    events_json,
-    allow_foreign_binding,
-    emit_result=True,
-    on_event=None,
-):
-    event_index = 0
-
-    def event(stage, status, text, root=None, activity=None):
-        nonlocal event_index
-        event_index += 1
-        payload = {
-            "schema": "kungfu.work-start.event/v1",
-            "index": event_index,
-            "stage": stage,
-            "status": status,
-            "text": text,
-            "root": root,
-        }
-        if activity is not None:
-            payload["activity"] = dict(activity)
-        if on_event is not None:
-            on_event(payload)
-        if events_json:
-            click.echo(json.dumps(payload, sort_keys=True))
-            click.get_text_stream("stdout").flush()
-
-    services = assignment_start.StartServices(
-        plan=_work_start_plan,
-        receipt=_work_start_receipt,
-        status=_status,
-        admit=_admit_captured_assignment,
-        admission_summary=_admission_summary,
-        profile_action=_profile_action,
-        claim_summary=_claim_summary,
-        advance_bound=_advance,
-        kickoff_summary=_kickoff_summary,
-        project_prompt=_project_work_prompt,
-        agent_report_summary=_agent_report_summary,
-    )
-    request = assignment_start.StartRequest(
-        config_home=ctx.config_home,
-        runtime_home=ctx.home,
-        request_file=request_file,
-        workspace_root=workspace_root,
-        home=home,
-        initiative_id=initiative_id,
-        assignment_id=assignment_id,
-        profile_id=profile_id,
-        actor=actor,
-        expected_plan_root=expected_plan_root,
-        execute=execute,
-        allow_foreign_binding=allow_foreign_binding,
-    )
-    result = assignment_start.execute(request, services, event)
-    if emit_result:
-        if events_json:
-            click.echo(json.dumps(result, sort_keys=True))
-        else:
-            _emit(result)
-    return result
-
-
-@assignment.command(
-    name="resume-prepare",
-    help="reconcile the exact Work Control Profile before product resume",
-    hidden=True,
-)
-@click.option("--workspace", "workspace_root", type=click.Path(file_okay=False))
-@click.option("--home", is_flag=True)
-@click.option("--actor", default="kungfu-product", show_default=True)
-@click.option("--execute", is_flag=True)
-@assignment_context
-def resume_prepare(
-    ctx,
-    workspace_root,
-    home,
-    actor,
-    execute,
-):
-    del ctx
-
-    def operation():
-        if not execute:
-            raise ValueError("resume-prepare requires explicit --execute")
-        identity, runtime_dir, _ = _runtime(
-            workspace_root,
-            home,
-            "read-only",
-        )
-        return {
-            **_prepare_resume_profile(runtime_dir, actor),
-            "workspace": identity.as_dict(),
-        }
-
-    _emit(_run(operation))
-
-
-@assignment.command(
-    name="start-resume",
-    help="resume the latest retained Starter Agent run without writing",
-)
-@click.option("--workspace", "workspace_root", type=click.Path(file_okay=False))
-@click.option("--home", is_flag=True)
-@click.option("--initiative-id", required=True)
-@click.option("--assignment-id", required=True)
-@assignment_context
-@surface(id="kungfu.work.start.resume")
-def start_resume(
-    ctx,
-    workspace_root,
-    home,
-    initiative_id,
-    assignment_id,
-):
-    del ctx
-    _emit(
-        _run(
-            lambda: assignment_evidence.resume_starter_work(
-                workspace_root=workspace_root,
-                home=home,
-                initiative_id=initiative_id,
-                assignment_id=assignment_id,
-                services=_EVIDENCE_SERVICES,
-            )
-        )
-    )
-
-
-def _reviewer_read_only_safe(reviewer):
-    return reviewer.get("provider") == "codex" or (
-        reviewer.get("provider") == "synthetic"
-        and reviewer.get("id") == "kungfu.mock-agent.review-fit"
-        and reviewer.get("source") == "qualification"
-    )
-
-
-def _work_review_plan(
-    *,
-    config_home,
-    runtime_home,
-    agent_report_file,
-    workspace_root,
-    home,
-    initiative_id,
-    assignment_id,
-    reviewer_profile_id,
-    allow_foreign_binding,
-):
-    identity, runtime_dir, _ = _runtime(workspace_root, home, "read-only")
-    status_value = _status(runtime_dir, initiative_id, assignment_id)
-    if status_value["phase"] not in {
-        "executing",
-        "stage-ready",
-        "completion-claimed",
-    }:
-        raise ValueError(
-            "independent review settlement requires executing, stage-ready, or "
-            f"completion-claimed Work, got {status_value['phase']}"
-        )
-    report_path, execution_report = assignment_evidence.load_execution_agent_report(
-        agent_report_file,
-        runtime_dir,
-        initiative_id,
-        assignment_id,
-    )
-    reviewer, selection = run_agent.select_profile(
-        reviewer_profile_id,
-        config_home=config_home,
-        runtime_home=runtime_home,
-    )
-    verification = run_agent.runtime_profiles.verify_profile(reviewer)
-    binding = orchestration.binding_provenance(allow_foreign=allow_foreign_binding)
-    workspace = Path(identity.workspace_root or "").resolve()
-    work_definition = status_value["assignment"]["work_definition"]
-    evidence = assignment_evidence.project_review_evidence(
-        workspace,
-        report_path,
-        work_definition,
-    )
-    acceptance_checks = list(work_definition.get("acceptance_criteria") or [])
-    if not acceptance_checks:
-        raise ValueError("Assignment has no acceptance criteria to review")
-    stable_verification = {
-        "ok": verification["ok"],
-        "available": verification["available"],
-        "version": verification["version"],
-        "error": verification["error"],
-    }
-    reviewer_safe = _reviewer_read_only_safe(reviewer)
-    body = {
-        "schema": "kungfu.work-review.plan/v1",
-        "workspace": {
-            "id": identity.workspace_id,
-            "root": str(workspace),
-            "identityRoot": identity.identity_root,
-        },
-        "work": {
-            "initiativeId": initiative_id,
-            "assignmentId": assignment_id,
-            "phase": status_value["phase"],
-            "queryProofRoot": status_value["query_proof_root"],
-            "assignmentRoot": initiative_family.semantic_root(
-                status_value["assignment"]
-            ),
-            "workDefinitionRoot": status_value["assignment"]["work_definition_root"],
-            "acceptanceChecks": acceptance_checks,
-        },
-        "deliverable": evidence["primary"],
-        "inputs": evidence["supporting"],
-        "evidenceMode": evidence["mode"],
-        "execution": {
-            "reportPath": str(report_path),
-            "reportRoot": execution_report["reportRoot"],
-            "runId": execution_report["runId"],
-            "episodeId": execution_report["episode"]["episodeId"],
-            "agent": execution_report["runtimeProfile"],
-            "workRef": execution_report["work"]["workRef"],
-        },
-        "reviewer": {
-            "id": reviewer["id"],
-            "label": reviewer["label"],
-            "provider": reviewer["provider"],
-            "profileRoot": run_agent.canonical_root(reviewer),
-            "selection": selection,
-            "verification": stable_verification,
-            "permissionMode": "read-only" if reviewer_safe else "unsupported",
-            "priorTranscriptBytes": 0,
-        },
-        "admissionBinding": {
-            "ok": binding["ok"],
-            "state": binding["state"],
-            "override": binding["override"],
-            "provenanceRoot": binding["provenance_root"],
-            "sourceRevision": binding["source_revision"],
-        },
-        "skippedEffects": [
-            "continuation-decision",
-            "git-commit",
-            "git-push",
-            "publication",
-        ],
-        "confirmationRequired": True,
-        "executable": bool(
-            binding["ok"] and stable_verification["ok"] and reviewer_safe
-        ),
-        "writeOccurred": False,
-    }
-    retained = _find_retained_reviewer_evidence(runtime_dir, body)
-    body["reviewExecution"] = (
-        {
-            "mode": "retained-evidence",
-            "reportPath": retained["reportPath"],
-            "reportRoot": retained["report"]["reportRoot"],
-            "runId": retained["report"]["runId"],
-            "episodeId": retained["report"]["episode"]["episodeId"],
-            "reviewCut": retained["report"]["work"]["workRef"]["systemTimeCut"],
-            "assessmentRoot": initiative_family.semantic_root(retained["assessment"]),
-        }
-        if retained is not None
-        else {
-            "mode": "fresh-process",
-            "reportPath": None,
-            "reportRoot": None,
-            "runId": None,
-            "episodeId": None,
-            "reviewCut": None,
-            "assessmentRoot": None,
-        }
-    )
-    body["reviewer"]["freshProcess"] = retained is None  # type: ignore[index]
-    effects = []
-    if retained is None:
-        effects.extend(
-            [
-                {
-                    "stage": "run",
-                    "label": "Launch one fresh reviewer in a read-only project sandbox",
-                },
-                {
-                    "stage": "assess",
-                    "label": "Check every acceptance criterion against retained sources",
-                },
-            ]
-        )
-    else:
-        effects.append(
-            {
-                "stage": "reuse",
-                "label": "Reuse the exact retained passing review without rerunning the Agent",
-            }
-        )
-    if status_value["phase"] == "executing":
-        effects.extend(
-            [
-                {
-                    "stage": "lease",
-                    "label": "Mint a new two-hour lease for bounded review settlement",
-                },
-                {
-                    "stage": "stage",
-                    "label": "Record stage-ready only when every criterion passes",
-                },
-            ]
-        )
-    if status_value["phase"] in {"executing", "stage-ready"}:
-        effects.append(
-            {
-                "stage": "claim",
-                "label": "Publish a proof-bound completion claim by local-user",
-            }
-        )
-    effects.append(
-        {
-            "stage": "review",
-            "label": "Record the reviewer as native independent review",
-        }
-    )
-    body["effects"] = effects
-    return {**body, "planRoot": initiative_family.semantic_root(body)}
-
-
 _find_retained_reviewer_evidence = assignment_review.find_passing_evidence
-
-
-def _mint_review_settlement_lease(runtime_dir, plan, actor):
-    lease_id = f"work-review-{uuid.uuid4().hex}"
-    lease_expires_at = (
-        (datetime.now(UTC) + timedelta(hours=2)).isoformat().replace("+00:00", "Z")
-    )
-    agent = plan["execution"]["agent"]
-    return _profile_action(
-        runtime_dir,
-        "claim-assignment",
-        {
-            "initiativeId": plan["work"]["initiativeId"],
-            "assignmentId": plan["work"]["assignmentId"],
-            "owner": actor,
-            "agent": agent["id"],
-            "slot": f"starter-{agent['provider']}",
-            "leaseId": lease_id,
-            "leaseExpiresAt": lease_expires_at,
-            "authorizedBy": actor,
-            "grantScope": "assignment-execution",
-            "actorType": "user",
-            "source": "kungfu",
-        },
-        actor,
-    )
-
-
-def _native_completion_review_values(plan, reviewer_report):
-    return {
-        "initiativeId": plan["work"]["initiativeId"],
-        "assignmentId": plan["work"]["assignmentId"],
-        "reviewer": plan["reviewer"]["id"],
-        "reviewerSource": reviewer_report["runId"],
-        "source": "kungfu",
-        "purpose": "handoff",
-        "cutSystemTime": 0,
-        # This selects how the native assessment itself executes. The external
-        # reviewer identity is carried separately by reviewer/reviewerSource.
-        "executorProfile": "thread",
-        "proposedFollowups": [],
-        # Starter Projects deliberately do not initialize Git. Their sealed
-        # Agent Episodes and proof roots are the completion evidence.
-        "checkoutPath": "",
-    }
-
-
-def _native_completion_claim_values(plan, reviewer_report, assessment):
-    return {
-        "initiativeId": plan["work"]["initiativeId"],
-        "assignmentId": plan["work"]["assignmentId"],
-        "statement": assessment["summary"],
-        "actor": "local-user",
-        "actorType": "user",
-        "source": "kungfu",
-        "evidenceEpisodeIds": [
-            int(plan["execution"]["episodeId"]),
-            int(reviewer_report["episode"]["episodeId"]),
-        ],
-        "assignmentSet": [plan["work"]["assignmentId"]],
-        "acceptanceRoot": plan["work"]["workDefinitionRoot"],
-        # Native Assignment evidence is carried by sealed Episodes and proof
-        # roots. Atlas and Project Cut roots belong only to legacy authority.
-        "proofRoots": [
-            plan["execution"]["reportRoot"],
-            reviewer_report["reportRoot"],
-            plan["deliverable"]["root"],
-        ],
-        "knownGaps": [],
-        "evidenceAvailability": [
-            {
-                "acceptance": criterion,
-                "level": "full",
-                "state": "available",
-            }
-            for criterion in plan["work"]["acceptanceChecks"]
-        ],
-    }
-
-
-@assignment.command(
-    name="review-agent-plan",
-    help="preview a fresh read-only Agent review and native completion review",
-)
-@click.argument(
-    "agent_report_file",
-    type=click.Path(exists=True, dir_okay=False, path_type=Path),
-)
-@click.option("--workspace", "workspace_root", type=click.Path(file_okay=False))
-@click.option("--home", is_flag=True)
-@click.option("--initiative-id", required=True)
-@click.option("--assignment-id", required=True)
-@click.option("--reviewer", "reviewer_profile_id", required=True)
-@click.option("--allow-foreign-binding", is_flag=True)
-@assignment_context
-def review_agent_plan(
-    ctx,
-    agent_report_file,
-    workspace_root,
-    home,
-    initiative_id,
-    assignment_id,
-    reviewer_profile_id,
-    allow_foreign_binding,
-):
-    _emit(
-        _run(
-            lambda: _work_review_plan(
-                config_home=ctx.config_home,
-                runtime_home=ctx.home,
-                agent_report_file=agent_report_file,
-                workspace_root=workspace_root,
-                home=home,
-                initiative_id=initiative_id,
-                assignment_id=assignment_id,
-                reviewer_profile_id=reviewer_profile_id,
-                allow_foreign_binding=allow_foreign_binding,
-            )
-        )
-    )
-
-
-@assignment.command(
-    name="review-agent-run",
-    help="run one exact fresh Agent review and submit passing native evidence",
-)
-@click.argument(
-    "agent_report_file",
-    type=click.Path(exists=True, dir_okay=False, path_type=Path),
-)
-@click.option("--workspace", "workspace_root", type=click.Path(file_okay=False))
-@click.option("--home", is_flag=True)
-@click.option("--initiative-id", required=True)
-@click.option("--assignment-id", required=True)
-@click.option("--reviewer", "reviewer_profile_id", required=True)
-@click.option("--expected-plan-root", required=True)
-@click.option("--execute", is_flag=True)
-@click.option("--events-json", is_flag=True)
-@click.option("--allow-foreign-binding", is_flag=True)
-@assignment_context
-@surface(id="kungfu.work.review.agent.run")
-def review_agent_run(
-    ctx,
-    agent_report_file,
-    workspace_root,
-    home,
-    initiative_id,
-    assignment_id,
-    reviewer_profile_id,
-    expected_plan_root,
-    execute,
-    events_json,
-    allow_foreign_binding,
-    emit_result=True,
-    on_event=None,
-):
-    event_index = 0
-
-    def event(stage, status, text, root=None, activity=None):
-        nonlocal event_index
-        event_index += 1
-        value = {
-            "schema": "kungfu.work-review.event/v1",
-            "index": event_index,
-            "stage": stage,
-            "status": status,
-            "text": text,
-            "root": root,
-        }
-        if activity is not None:
-            value["activity"] = activity
-        if on_event is not None:
-            on_event(value)
-        if events_json:
-            click.echo(json.dumps(value, sort_keys=True))
-            click.get_text_stream("stdout").flush()
-
-    services = assignment_review_lifecycle.ReviewServices(
-        plan=_work_review_plan,
-        receipt=_work_start_receipt,
-        runtime=_assignment_runtime,
-        retained_evidence=_find_retained_reviewer_evidence,
-        agent_report_summary=_agent_report_summary,
-        status=_status,
-        mint_lease=_mint_review_settlement_lease,
-        advance=_advance,
-        completion_claim_values=_native_completion_claim_values,
-        profile_action=_profile_action,
-        completion_review_values=_native_completion_review_values,
-    )
-    request = assignment_review_lifecycle.ReviewRequest(
-        config_home=ctx.config_home,
-        runtime_home=ctx.home,
-        agent_report_file=agent_report_file,
-        workspace_root=workspace_root,
-        home=home,
-        initiative_id=initiative_id,
-        assignment_id=assignment_id,
-        reviewer_profile_id=reviewer_profile_id,
-        expected_plan_root=expected_plan_root,
-        execute=execute,
-        allow_foreign_binding=allow_foreign_binding,
-    )
-    result = assignment_review_lifecycle.execute(
-        request, services, event, lambda: event_index
-    )
-    if emit_result:
-        if events_json:
-            click.echo(json.dumps(result, sort_keys=True))
-        else:
-            _emit(result)
-    return result
 
 
 @assignment.command(
@@ -1486,176 +871,6 @@ def status(ctx, workspace_root, home, initiative_id, assignment_id, now):
     _emit(_run(operation))
 
 
-@assignment.command(
-    name="family-contract",
-    help="show the versioned native Initiative-family protocol",
-)
-@assignment_context
-def family_contract_command(ctx):
-    _emit(initiative_family.family_contract())
-
-
-@assignment.command(
-    name="family-create",
-    help="create one rooted inert-parent and bounded-Wave family state",
-)
-@click.argument(
-    "blueprint_file",
-    type=click.Path(exists=True, dir_okay=False, path_type=Path),
-)
-@click.option("--out", type=click.Path(dir_okay=False, path_type=Path))
-@assignment_context
-def family_create(ctx, blueprint_file, out):
-    def operation():
-        blueprint = json.loads(blueprint_file.read_text(encoding="utf-8"))
-        state = initiative_family.create_family_state(blueprint)
-        return {
-            "schema": "kungfu.work-control.initiative-family-create/v1",
-            "state": state,
-            "stateRoot": state["stateRoot"],
-            "outputPath": _write_immutable_json(out, state),
-            "verification": initiative_family.verify_family_state(state),
-        }
-
-    _emit(_run(operation))
-
-
-@assignment.command(
-    name="family-transition",
-    help="append one expected-root terminal or acceptance transition",
-)
-@click.argument(
-    "state_file",
-    type=click.Path(exists=True, dir_okay=False, path_type=Path),
-)
-@click.argument(
-    "transition_file",
-    type=click.Path(exists=True, dir_okay=False, path_type=Path),
-)
-@click.option("--out", type=click.Path(dir_okay=False, path_type=Path))
-@assignment_context
-def family_transition(ctx, state_file, transition_file, out):
-    def operation():
-        state = json.loads(state_file.read_text(encoding="utf-8"))
-        transition = json.loads(transition_file.read_text(encoding="utf-8"))
-        successor = initiative_family.transition_family_state(state, transition)
-        return {
-            "schema": "kungfu.work-control.initiative-family-transition-result/v1",
-            "state": successor,
-            "stateRoot": successor["stateRoot"],
-            "previousStateRoot": successor["previousStateRoot"],
-            "outputPath": _write_immutable_json(out, successor),
-            "verification": initiative_family.verify_family_state(successor),
-        }
-
-    _emit(_run(operation))
-
-
-@assignment.command(
-    name="family-verify",
-    help="verify one native Initiative-family state without runtime mutation",
-)
-@click.argument(
-    "state_file",
-    type=click.Path(exists=True, dir_okay=False, path_type=Path),
-)
-@assignment_context
-def family_verify(ctx, state_file):
-    def operation():
-        state = json.loads(state_file.read_text(encoding="utf-8"))
-        return initiative_family.verify_family_state(state)
-
-    _emit(_run(operation))
-
-
-@assignment.command(
-    name="family-contract-v2",
-    help="show the additive typed Initiative-family envelope protocol",
-)
-@assignment_context
-def family_contract_v2_command(ctx):
-    _emit(initiative_family_v2.family_contract_v2())
-
-
-@assignment.command(
-    name="family-upgrade-v2",
-    help="explicitly bind one immutable v1 state into a typed v2 envelope",
-)
-@click.argument(
-    "state_file",
-    type=click.Path(exists=True, dir_okay=False, path_type=Path),
-)
-@click.argument(
-    "binding_file",
-    type=click.Path(exists=True, dir_okay=False, path_type=Path),
-)
-@click.option("--out", type=click.Path(dir_okay=False, path_type=Path))
-@assignment_context
-def family_upgrade_v2(ctx, state_file, binding_file, out):
-    def operation():
-        state = json.loads(state_file.read_text(encoding="utf-8"))
-        bindings = json.loads(binding_file.read_text(encoding="utf-8"))
-        upgrade = initiative_family_v2.upgrade_family_state_v2(state, bindings)
-        successor = upgrade["successorState"]
-        return {
-            **upgrade,
-            "outputPath": _write_immutable_json(out, successor),
-            "verification": initiative_family_v2.verify_family_state_v2(successor),
-        }
-
-    _emit(_run(operation))
-
-
-@assignment.command(
-    name="family-transition-v2",
-    help="advance a typed family state with an exact v1 transition and bindings",
-)
-@click.argument(
-    "state_file",
-    type=click.Path(exists=True, dir_okay=False, path_type=Path),
-)
-@click.argument(
-    "transition_file",
-    type=click.Path(exists=True, dir_okay=False, path_type=Path),
-)
-@click.option("--out", type=click.Path(dir_okay=False, path_type=Path))
-@assignment_context
-def family_transition_v2(ctx, state_file, transition_file, out):
-    def operation():
-        state = json.loads(state_file.read_text(encoding="utf-8"))
-        transition = json.loads(transition_file.read_text(encoding="utf-8"))
-        successor = initiative_family_v2.transition_family_state_v2(state, transition)
-        return {
-            "schema": "kungfu.work-control.initiative-family-transition-result/v2",
-            "state": successor,
-            "stateRoot": successor["stateRoot"],
-            "previousStateRoot": successor["previousStateRoot"],
-            "v1ProjectionRoot": successor["v1ProjectionRoot"],
-            "typedBindingRoot": successor["typedBindingRoot"],
-            "outputPath": _write_immutable_json(out, successor),
-            "verification": initiative_family_v2.verify_family_state_v2(successor),
-        }
-
-    _emit(_run(operation))
-
-
-@assignment.command(
-    name="family-verify-v2",
-    help="read v1 as under-typed or verify one complete typed v2 state",
-)
-@click.argument(
-    "state_file",
-    type=click.Path(exists=True, dir_okay=False, path_type=Path),
-)
-@assignment_context
-def family_verify_v2(ctx, state_file):
-    def operation():
-        state = json.loads(state_file.read_text(encoding="utf-8"))
-        return initiative_family_v2.verify_family_state_v2(state)
-
-    _emit(_run(operation))
-
-
 @assignment.command(help="evaluate the native run or closeout gate")
 @assignment_identity_options
 @click.option("--target", type=click.Choice(["run", "closeout"]), required=True)
@@ -1735,193 +950,33 @@ review = _json_action("review", "review-completion")
 decide = _json_action("decide", "decide-continuation")
 
 
-@assignment.command(
-    name="binding-create",
-    help="build one path-free cross-workspace parent/child binding from receipts",
+from kungfu.cli.commands._assignment.start import (  # noqa: E402
+    start_plan as start_plan,
+    start_work as start_work,
+    resume_prepare as resume_prepare,
+    start_resume as start_resume,
 )
-@click.option(
-    "--parent-admission",
-    type=click.Path(exists=True, dir_okay=False, path_type=Path),
-    required=True,
+from kungfu.cli.commands._assignment.review import (  # noqa: E402
+    _reviewer_read_only_safe as _reviewer_read_only_safe,
+    _work_review_plan as _work_review_plan,
+    _mint_review_settlement_lease as _mint_review_settlement_lease,
+    _native_completion_review_values as _native_completion_review_values,
+    _native_completion_claim_values as _native_completion_claim_values,
+    review_agent_plan as review_agent_plan,
+    review_agent_run as review_agent_run,
 )
-@click.option(
-    "--parent-status",
-    type=click.Path(exists=True, dir_okay=False, path_type=Path),
-    required=True,
+from kungfu.cli.commands._assignment.family import (  # noqa: E402
+    family_contract_command as family_contract_command,
+    family_create as family_create,
+    family_transition as family_transition,
+    family_verify as family_verify,
+    family_contract_v2_command as family_contract_v2_command,
+    family_upgrade_v2 as family_upgrade_v2,
+    family_transition_v2 as family_transition_v2,
+    family_verify_v2 as family_verify_v2,
+    binding_create as binding_create,
+    bind as bind,
+    verify_binding as verify_binding,
+    seal as seal,
+    verify_seal as verify_seal,
 )
-@click.option(
-    "--child-admission",
-    type=click.Path(exists=True, dir_okay=False, path_type=Path),
-    required=True,
-)
-@click.option(
-    "--child-status",
-    type=click.Path(exists=True, dir_okay=False, path_type=Path),
-    required=True,
-)
-@click.option("--out", type=click.Path(dir_okay=False, path_type=Path))
-@assignment_context
-def binding_create(
-    ctx,
-    parent_admission,
-    parent_status,
-    child_admission,
-    child_status,
-    out,
-):
-    def binding_create_operation():
-        binding = orchestration.cross_workspace_binding(
-            json.loads(parent_admission.read_text(encoding="utf-8")),
-            json.loads(parent_status.read_text(encoding="utf-8")),
-            json.loads(child_admission.read_text(encoding="utf-8")),
-            json.loads(child_status.read_text(encoding="utf-8")),
-        )
-        output_path = None
-        if out is not None:
-            output_path = out.expanduser().resolve()
-            content = (initiative_family.canonical_json(binding) + "\n").encode("utf-8")
-            if output_path.exists() and output_path.read_bytes() != content:
-                raise ValueError("binding output exists with different bytes")
-            output_path.parent.mkdir(parents=True, exist_ok=True)
-            if not output_path.exists():
-                output_path.write_bytes(content)
-        return {
-            "schema": "kungfu.assignment-orchestration.cross-workspace-binding-create/v1",
-            "binding": binding,
-            "bindingRoot": binding["bindingRoot"],
-            "outputPath": str(output_path) if output_path else None,
-            "next_actions": [
-                {
-                    "action": "bind",
-                    "description": "admit the exact binding in both endpoint workspaces",
-                }
-            ],
-        }
-
-    _emit(_run(binding_create_operation))
-
-
-@assignment.command(
-    name="bind",
-    help="plan or admit one exact cross-workspace binding in this endpoint",
-)
-@click.argument(
-    "binding_file", type=click.Path(exists=True, dir_okay=False, path_type=Path)
-)
-@assignment_identity_options
-@click.option("--execute", is_flag=True)
-@click.option("--expected-binding-root", default="")
-@assignment_context
-def bind(
-    ctx,
-    binding_file,
-    workspace_root,
-    home,
-    initiative_id,
-    assignment_id,
-    execute,
-    expected_binding_root,
-):
-    def bind_operation():
-        binding = json.loads(binding_file.read_text(encoding="utf-8"))
-        identity, runtime_dir, _ = _runtime(workspace_root, home, "read-only")
-        current = _status(runtime_dir, initiative_id, assignment_id)
-        plan = orchestration.cross_workspace_binding_plan(
-            identity.workspace_root or identity.data_home,
-            identity.as_dict(),
-            current,
-            binding,
-        )
-        if not execute:
-            return {
-                **plan,
-                "next_actions": [
-                    {
-                        "action": "bind",
-                        "expected_binding_root": plan["bindingRoot"],
-                    }
-                ],
-            }
-        return orchestration.apply_cross_workspace_binding(
-            plan, binding, expected_binding_root
-        )
-
-    _emit(_run(bind_operation))
-
-
-@assignment.command(
-    name="verify-binding",
-    help="verify a local cross-workspace binding receipt without a live runtime",
-)
-@click.option(
-    "--binding",
-    "binding_file",
-    type=click.Path(exists=True, dir_okay=False, path_type=Path),
-    required=True,
-)
-@click.option(
-    "--receipt",
-    "receipt_file",
-    type=click.Path(exists=True, dir_okay=False, path_type=Path),
-    required=True,
-)
-@assignment_context
-def verify_binding(ctx, binding_file, receipt_file):
-    result = _run(
-        lambda: orchestration.verify_cross_workspace_binding_receipt(
-            binding_file, receipt_file
-        )
-    )
-    _emit(result)
-    if not result["ok"]:
-        raise click.exceptions.Exit(5)
-
-
-@assignment.command(help="plan or write a portable content-addressed state snapshot")
-@assignment_identity_options
-@click.option("--execute", is_flag=True)
-@click.option("--expected-state-root", default="")
-@assignment_context
-def seal(
-    ctx,
-    workspace_root,
-    home,
-    initiative_id,
-    assignment_id,
-    execute,
-    expected_state_root,
-):
-    def seal_operation():
-        identity, runtime_dir, _ = _runtime(workspace_root, home)
-        _ensure_profile(runtime_dir, "assignment-seal")
-        current = _status(runtime_dir, initiative_id, assignment_id)
-        plan = orchestration.sealed_state_plan(
-            identity.workspace_root or identity.data_home,
-            current,
-            workspace_identity=identity.as_dict(),
-        )
-        if not execute:
-            return {
-                **plan,
-                "executed": False,
-                "next_actions": [
-                    {"action": "seal", "expected_state_root": plan["state_root"]}
-                ],
-            }
-        return orchestration.apply_sealed_state(plan, expected_state_root)
-
-    _emit(_run(seal_operation))
-
-
-@assignment.command(
-    name="verify-seal", help="verify sealed state without a live runtime"
-)
-@click.argument(
-    "state_file", type=click.Path(exists=True, dir_okay=False, path_type=Path)
-)
-@assignment_context
-def verify_seal(ctx, state_file):
-    result = _run(lambda: orchestration.verify_sealed_state(state_file))
-    _emit(result)
-    if not result["ok"]:
-        raise click.exceptions.Exit(5)
