+++ framework/core/src/python/kungfu/cli/commands/profile.py
@@ -1,6 +1,5 @@
 # SPDX-License-Identifier: Apache-2.0
 
-import base64
 import json
 from pathlib import Path
 
@@ -49,13 +48,6 @@ def _run(fn):
 
 def _load_json(path):
     return json.loads(Path(path).read_text(encoding="utf-8"))
-
-
-def _decode_json(value):
-    try:
-        return json.loads(base64.b64decode(value, validate=True).decode("utf-8"))
-    except (ValueError, UnicodeDecodeError, json.JSONDecodeError) as error:
-        raise click.UsageError("invalid base64 JSON input") from error
 
 
 @profile.command(help="show the installed Agent Profile SDK contract")
@@ -963,57 +955,7 @@ def assessment_authorize(ctx, plan_base64, choice, authorized_by, as_json):
     _json(_run(operation))
 
 
-@profile.command(help="plan or invoke a declarative Profile action")
-@click.argument("source", type=click.Path(exists=True, file_okay=False, path_type=Path))
-@click.argument("action_id")
-@click.option(
-    "--input",
-    "input_path",
-    type=click.Path(exists=True, dir_okay=False, path_type=Path),
+from kungfu.cli.commands._profile.action import (  # noqa: E402
+    _decode_json as _decode_json,
+    invoke as invoke,
 )
-@click.option(
-    "--plan-file", type=click.Path(exists=True, dir_okay=False, path_type=Path)
-)
-@click.option(
-    "--authorization-file",
-    type=click.Path(exists=True, dir_okay=False, path_type=Path),
-)
-@click.option("--out", type=click.Path(dir_okay=False, path_type=Path))
-@click.option("--execute", is_flag=True)
-@click.option("--json", "as_json", is_flag=True)
-@profile_context
-def invoke(
-    ctx,
-    source,
-    action_id,
-    input_path,
-    plan_file,
-    authorization_file,
-    out,
-    execute,
-    as_json,
-):
-    input_value = _load_json(input_path) if input_path else {}
-    if execute:
-        if plan_file is None:
-            raise click.UsageError("--execute requires --plan-file")
-        payload = _run(
-            lambda: profile_sdk.authorized_action_invoke(
-                ctx.runtime_dir,
-                _load_json(plan_file),
-                _load_json(authorization_file) if authorization_file else None,
-            )
-        )
-    else:
-        payload = _run(
-            lambda: profile_sdk.plan_action(
-                source, ctx.runtime_dir, action_id, input_value
-            )
-        )
-        if out:
-            out.write_text(
-                json.dumps(payload, indent=2, sort_keys=True) + "\n",
-                encoding="utf-8",
-            )
-            payload["actionPlanPath"] = str(out.resolve())
-    _json(payload)
