+++ framework/core/src/python/kungfu/profile_sdk.py
@@ -15,7 +15,6 @@ mutation to the Core service introduced by KF-ADR-019f86da-4f90-7f46-b195-3af622
 from __future__ import annotations
 
 import hashlib
-import base64
 import copy
 import importlib.util
 import json
@@ -30,7 +29,7 @@ from pathlib import Path
 from typing import Any, Iterator, Mapping
 
 from kungfu import agent as agent_pack
-from kungfu import runtime_broker
+from kungfu import runtime_broker  # noqa: F401 -- private owner seam
 from kungfu import kfx_contract
 from kungfu.storage import profile_member_state
 from kungfu.storage import service as storage_service
@@ -75,7 +74,7 @@ from kungfu.profile_sdk_source import (
     package_content_root as package_content_root,
     resolve_profile_source as _resolve_profile_source,
 )
-from kungfu.profile_sdk_kfd3 import (
+from kungfu.profile_sdk_kfd3 import (  # noqa: F401 -- private owner seams
     Kfd3Operations as Kfd3Operations,
     _agent_interface_authority as _agent_interface_authority,
     _earn_kfd3 as _earn_kfd3_impl,
@@ -855,1162 +854,75 @@ def application(
     }
 
 
-def intent_inspect(
-    source: str | Path, runtime_dir: str | Path, intent_id: str
-) -> dict[str, Any]:
-    projection = application(source, runtime_dir, include_qualification=False)
-    intent = next(
-        (row for row in projection["intents"] if row["id"] == intent_id), None
-    )
-    if intent is None:
-        raise ProfileSdkError(
-            "collaboration-intent-not-found", f"intent not found: {intent_id}"
-        )
-    return {
-        "schema": "kungfu.profile-intent-inspection/v1",
-        "profileId": projection["profileId"],
-        "profileSuiteRoot": projection["profileSuiteRoot"],
-        "collaborationRoot": projection["collaborationRoot"],
-        "closureRoot": projection["closureRoot"],
-        "source": projection["source"],
-        "profileRevision": projection["profileRevision"],
-        "activeExactRoot": projection["activeExactRoot"],
-        "intent": intent,
-        "cut": {
-            "kind": "profile-revision",
-            "value": projection["profileRevision"],
-        },
-    }
-
-
-def intent_advise(
-    source: str | Path, runtime_dir: str | Path, intent_id: str
-) -> dict[str, Any]:
-    inspected = intent_inspect(source, runtime_dir, intent_id)
-    intent = inspected["intent"]
-    application_value = application(source, runtime_dir, include_qualification=False)
-    constraints = [
-        row
-        for row in application_value["constraints"]
-        if "*" in row["appliesTo"] or intent_id in row["appliesTo"]
-    ]
-    eligible = inspected["activeExactRoot"] and not intent["missingCapabilities"]
-    return {
-        "schema": "kungfu.profile-intent-advice/v1",
-        "profileId": inspected["profileId"],
-        "profileSuiteRoot": inspected["profileSuiteRoot"],
-        "collaborationRoot": inspected["collaborationRoot"],
-        "closureRoot": inspected["closureRoot"],
-        "source": inspected["source"],
-        "intentId": intent_id,
-        "eligible": eligible,
-        "recommendation": "preview" if eligible else "resolve-preconditions",
-        "constraints": constraints,
-        "knownLimits": application_value["knownLimits"],
-        "missingCapabilities": intent["missingCapabilities"],
-        "preconditions": {
-            "activeExactRoot": inspected["activeExactRoot"],
-            "profileRevision": inspected["profileRevision"],
-        },
-    }
-
-
-def intent_plan(
-    source: str | Path,
-    runtime_dir: str | Path,
-    intent_id: str,
-    input_value: Any,
-) -> dict[str, Any]:
-    advice = intent_advise(source, runtime_dir, intent_id)
-    if not advice["eligible"]:
-        raise ProfileSdkError(
-            "intent-precondition-failed",
-            "intent cannot be previewed until its active-root and capability preconditions hold",
-            advice=advice,
-        )
-    inspected = intent_inspect(source, runtime_dir, intent_id)
-    action_plan = plan_action(
-        source, runtime_dir, inspected["intent"]["actionId"], input_value
-    )
-    identity = {
-        "profileSuiteRoot": inspected["profileSuiteRoot"],
-        "collaborationRoot": inspected["collaborationRoot"],
-        "closureRoot": inspected["closureRoot"],
-        "intentId": intent_id,
-        "actionPlanId": action_plan["planId"],
-        "input": input_value,
-    }
-    return {
-        "schema": INTENT_PLAN_SCHEMA,
-        "planId": _root(identity),
-        **identity,
-        "source": inspected["source"],
-        "actionPlan": action_plan,
-        "decisionCard": action_plan.get("decisionCard"),
-        "protocolStage": "preview",
-    }
-
-
-def intent_apply(
-    runtime_dir: str | Path,
-    plan: Mapping[str, Any],
-    answer: Mapping[str, Any] | None,
-) -> dict[str, Any]:
-    if plan.get("schema") != INTENT_PLAN_SCHEMA:
-        raise ProfileSdkError(
-            "intent-plan-invalid", "intent apply requires an intent plan"
-        )
-    refreshed = intent_plan(
-        str(plan.get("source") or ""),
-        runtime_dir,
-        str(plan.get("intentId") or ""),
-        plan.get("input"),
-    )
-    if refreshed["planId"] != plan.get("planId"):
-        raise ProfileSdkError("intent-plan-stale", "intent plan changed after preview")
-    action_receipt = authorized_action_invoke(
-        runtime_dir, plan.get("actionPlan") or {}, answer
-    )
-    identity = {
-        "planId": plan["planId"],
-        "actionPlanId": plan["actionPlanId"],
-        "profileSuiteRoot": plan["profileSuiteRoot"],
-        "collaborationRoot": plan["collaborationRoot"],
-        "closureRoot": plan["closureRoot"],
-        "intentId": plan["intentId"],
-        "actionReceipt": action_receipt,
-    }
-    return {
-        "schema": INTENT_RECEIPT_SCHEMA,
-        "receiptId": _root(identity),
-        **identity,
-        "source": plan["source"],
-        "protocolStage": "receipt",
-        "executionReceiptVerified": action_receipt["verified"],
-        "verified": False,
-    }
-
-
-def intent_verify(
-    source: str | Path, runtime_dir: str | Path, receipt: Mapping[str, Any]
-) -> dict[str, Any]:
-    if receipt.get("schema") != INTENT_RECEIPT_SCHEMA:
-        raise ProfileSdkError(
-            "intent-receipt-invalid", "verify requires an intent receipt"
-        )
-    inspected = intent_inspect(source, runtime_dir, str(receipt.get("intentId") or ""))
-    current = {
-        "profileSuiteRoot": inspected["profileSuiteRoot"],
-        "collaborationRoot": inspected["collaborationRoot"],
-        "closureRoot": inspected["closureRoot"],
-    }
-    expected = {key: receipt.get(key) for key in current}
-    if current != expected:
-        raise ProfileSdkError(
-            "intent-receipt-stale",
-            "Profile or collaboration closure changed after execution",
-            expected=expected,
-            actual=current,
-        )
-    action_receipt = receipt.get("actionReceipt") or {}
-    verified = bool(
-        receipt.get("executionReceiptVerified")
-        and action_receipt.get("verified")
-        and action_receipt.get("planId") == receipt.get("actionPlanId")
-    )
-    if not verified:
-        raise ProfileSdkError(
-            "intent-execution-unverified", "underlying action receipt is not verified"
-        )
-    return {
-        "schema": "kungfu.profile-intent-verification/v1",
-        "receiptId": receipt["receiptId"],
-        **current,
-        "intentId": receipt["intentId"],
-        "verifyView": inspected["intent"]["verifyView"],
-        "protocolStage": "verify",
-        "verified": True,
-        "evidenceScope": "profile-root/collaboration-root/execution-receipt/declared-verify-view",
-        "knownLimit": "declared verify view is bound but domain outcome truth remains evidence-dependent",
-    }
-
-
-def authorize_current_intent(
-    runtime_dir: str | Path,
-    source: str | Path,
-    intent_id: str,
-    input_value: Any,
-    expected_plan_id: str,
-    choice: str,
-    authorized_by: str,
-) -> dict[str, Any]:
-    plan = intent_plan(source, runtime_dir, intent_id, input_value)
-    if plan["planId"] != expected_plan_id:
-        raise ProfileSdkError(
-            "intent-plan-stale",
-            "intent plan changed; review a fresh preview",
-            expectedPlanId=expected_plan_id,
-            actualPlanId=plan["planId"],
-        )
-    answer = answer_decision(plan["decisionCard"], choice, authorized_by)
-    receipt = intent_apply(runtime_dir, plan, answer)
-    return {**receipt, "verification": intent_verify(source, runtime_dir, receipt)}
-
-
-def _kfd3_operations() -> Kfd3Operations:
-    return Kfd3Operations(
-        validate_source=validate_source,
-        application=application,
-        intent_plan=intent_plan,
-        decision_card=decision_card,
-        lifecycle_plan=lifecycle_plan,
-        lifecycle_apply=lifecycle_apply,
-        answer_decision=answer_decision,
-    )
-
-
-def _earn_kfd3(
-    source: str | Path,
-    runtime_dir: str | Path,
-    *,
-    qualification_source: str,
-) -> dict[str, Any]:
-    return _earn_kfd3_impl(
-        _kfd3_operations(),
-        source,
-        runtime_dir,
-        qualification_source=qualification_source,
-    )
-
-
-def build_kfd3_release_manifest(
-    sources: list[str | Path], runtime_dir: str | Path
-) -> dict[str, Any]:
-    return _build_kfd3_release_manifest(_kfd3_operations(), sources, runtime_dir)
-
-
-def kfd3_status(source: str | Path, runtime_dir: str | Path) -> dict[str, Any]:
-    return _kfd3_status(_kfd3_operations(), source, runtime_dir)
-
-
-def kfd3_qualification_plan(
-    source: str | Path, runtime_dir: str | Path
-) -> dict[str, Any]:
-    return _kfd3_qualification_plan(_kfd3_operations(), source, runtime_dir)
-
-
-def qualify_kfd3(
-    source: str | Path,
-    runtime_dir: str | Path,
-    *,
-    authorization_id: str = "kfd3-cli-explicit",
-    qualification_source: str = "local",
-) -> dict[str, Any]:
-    return _qualify_kfd3(
-        _kfd3_operations(),
-        source,
-        runtime_dir,
-        authorization_id=authorization_id,
-        qualification_source=qualification_source,
-    )
-
-
-def authorize_kfd3_qualification(
-    source: str | Path,
-    runtime_dir: str | Path,
-    expected_plan_id: str,
-    choice: str,
-    authorized_by: str,
-) -> dict[str, Any]:
-    return _authorize_kfd3_qualification(
-        _kfd3_operations(),
-        source,
-        runtime_dir,
-        expected_plan_id,
-        choice,
-        authorized_by,
-    )
-
-
-def verify_kfd3(
-    source: str | Path, runtime_dir: str | Path, receipt: Mapping[str, Any]
-) -> dict[str, Any]:
-    return _verify_kfd3(_kfd3_operations(), source, runtime_dir, receipt)
-
-
-def qualify_source(source: str | Path, runtime_dir: str | Path) -> dict[str, Any]:
-    validated = validate_source(source, runtime_dir)
-    inspection = validated["inspection"]
-    compatibility = _read_ref_json(
-        inspection, inspection["profile"]["kfd1"]["compatibility"]
-    )
-    qualification = _read_ref_json(
-        inspection, inspection["profile"]["qualification"]["profile"]
-    )
-    contracts = compatibility.get("runtimeContracts", [])
-    checks = qualification.get("checks", [])
-    if (
-        contracts != ["kungfu.profile-lifecycle/v1"]
-        and "kungfu.profile-lifecycle/v1" not in contracts
-    ):
-        raise ProfileSdkError(
-            "runtime-incompatible", "Profile omits the current lifecycle contract"
-        )
-    if sorted(checks) != ["content-closure", "runtime-contract"]:
-        raise ProfileSdkError(
-            "qualification-check-unsupported",
-            "This runtime only qualifies content-closure and runtime-contract",
-            requested=checks,
-        )
-    collaboration = validated["collaboration"]
-    return {
-        "schema": "kungfu.profile-source-qualification/v1",
-        "profileSuiteRoot": inspection["profile_suite_root"],
-        "status": "qualified-for-install-plan",
-        "checks": sorted(checks),
-        "evidenceScope": "source-contract/content-closure/runtime-contract",
-        "kfd3": {
-            "declared": collaboration["declared"],
-            "qualified": False,
-            "status": collaboration.get("qualificationStatus", "not-declared"),
-            "closureRoot": collaboration.get("closureRoot"),
-            "reason": collaboration.get("reason"),
-        },
-        "workConformance": _work_profile_conformance(inspection, "qualify"),
-        "lifecycleMutation": False,
-    }
-
-
-def export_source_bundle(
-    source: str | Path, runtime_dir: str | Path, *, thin: bool = False
-) -> dict[str, Any]:
-    """Export an exact Profile source closure without lifecycle side effects."""
-
-    validated = validate_source(source, runtime_dir)
-    root = Path(validated["source"]["source"]).resolve()
-    entries = []
-    for path in _portable_source_paths(root):
-        data = path.read_bytes()
-        entry = {
-            "path": path.relative_to(root).as_posix(),
-            "sha256": _sha256(data),
-            "size": len(data),
-        }
-        if not thin:
-            entry["contentBase64"] = base64.b64encode(data).decode("ascii")
-        entries.append(entry)
-    body = {
-        "schema": SOURCE_BUNDLE_SCHEMA,
-        "mode": "thin" if thin else "full",
-        "selfContained": not thin,
-        "profileId": validated["source"]["profile"]["id"],
-        "profileVersion": validated["source"]["profile"]["version"],
-        "profileSuiteRoot": validated["inspection"]["profile_suite_root"],
-        "memberRoots": validated["source"]["memberRoots"],
-        "entries": entries,
-    }
-    body["bundleRoot"] = _root(body)
-    _validate_sdk_value("sourceBundleSchema", body, "Profile source bundle")
-    return body
-
-
-def source_import_plan(
-    bundle: Mapping[str, Any], destination: str | Path
-) -> dict[str, Any]:
-    """Plan reconstruction of source bytes; never install or activate a Profile."""
-
-    normalized = _validate_source_bundle(bundle)
-    target = Path(destination).expanduser().resolve()
-    collision = target.exists() and (not target.is_dir() or any(target.iterdir()))
-    identity = {
-        "bundle": normalized,
-        "destination": str(target),
-        "destinationCollision": collision,
-    }
-    plan = {
-        "schema": SOURCE_IMPORT_PLAN_SCHEMA,
-        "planId": _root(identity),
-        **identity,
-        "requiresAuthorization": normalized["mode"] == "full" and not collision,
-    }
-    if normalized["mode"] == "thin":
-        plan["decisionCard"] = decision_card(
-            "profile-source-material-required",
-            "Thin Profile bundles are root inventories; supply a full bundle before reconstruction.",
-            choices=["supply-full-bundle"],
-            basis={"bundleRoot": normalized["bundleRoot"]},
-            required_authority="profile-source-owner",
-            resume_command="export or obtain the exact full Profile source bundle",
-        )
-    elif collision:
-        plan["decisionCard"] = decision_card(
-            "profile-source-collision",
-            "The Profile source import destination is not empty.",
-            choices=["choose-an-empty-directory", "inspect-and-merge-manually"],
-            basis={"bundleRoot": normalized["bundleRoot"], "destination": str(target)},
-            required_authority="profile-source-owner",
-            resume_command="choose an empty destination and re-plan",
-        )
-    else:
-        plan["decisionCard"] = decision_card(
-            "profile-source-import-authorization",
-            "Authorize reconstruction of this exact Profile source closure without lifecycle activation.",
-            choices=["approve", "deny"],
-            basis={"planId": plan["planId"], "bundleRoot": normalized["bundleRoot"]},
-            required_authority="profile-source-owner",
-            resume_command="answer this card, then apply the exact source import plan",
-        )
-    return plan
-
-
-def authorized_source_import(
-    plan: Mapping[str, Any], answer: Mapping[str, Any]
-) -> dict[str, Any]:
-    if plan.get("schema") != SOURCE_IMPORT_PLAN_SCHEMA:
-        raise ProfileSdkError(
-            "source-import-plan-invalid", "Profile source import requires an exact plan"
-        )
-    refreshed = source_import_plan(
-        dict(plan.get("bundle") or {}), str(plan.get("destination") or "")
-    )
-    if refreshed["planId"] != plan.get("planId"):
-        raise ProfileSdkError(
-            "source-import-plan-stale", "Profile source bundle or destination changed"
-        )
-    if not refreshed["requiresAuthorization"]:
-        raise ProfileSdkError(
-            "source-import-not-ready",
-            "Profile source import needs a full bundle and an empty destination",
-            decisionCards=[refreshed["decisionCard"]],
-        )
-    validate_decision_answer(answer, refreshed["decisionCard"])
-    if (
-        answer.get("choice") != "approve"
-        or (answer.get("basis") or {}).get("planId") != refreshed["planId"]
-    ):
-        raise ProfileSdkError(
-            "decision-denied", "Profile source import was not approved"
-        )
-    target = Path(refreshed["destination"])
-    target.mkdir(parents=True, exist_ok=True)
-    written = []
-    for entry in refreshed["bundle"]["entries"]:
-        path = _confined(target, entry["path"])
-        path.parent.mkdir(parents=True, exist_ok=True)
-        data = base64.b64decode(entry["contentBase64"], validate=True)
-        path.write_bytes(data)
-        written.append(entry["path"])
-    return {
-        "schema": "kungfu.profile-source-import-receipt/v1",
-        "planId": refreshed["planId"],
-        "authorizationId": answer["authorizationId"],
-        "bundleRoot": refreshed["bundle"]["bundleRoot"],
-        "profileSuiteRoot": refreshed["bundle"]["profileSuiteRoot"],
-        "destination": str(target),
-        "written": written,
-        "lifecycleMutation": False,
-    }
-
-
-def lifecycle_plan(
-    runtime_dir: str | Path,
-    action: str,
-    source: str | Path | None = None,
-    **values: Any,
-) -> dict[str, Any]:
-    request: dict[str, Any] = {"action": action, **values}
-    if source is not None:
-        resolved = resolve_source(source)
-        request.update(
-            profile_path=resolved["profilePath"],
-            member_roots=resolved["memberRoots"],
-        )
-        if action in {"qualify", "activate"}:
-            inspection = storage_service.profile_lifecycle(
-                runtime_dir,
-                "inspect",
-                profile_path=resolved["profilePath"],
-                member_roots=resolved["memberRoots"],
-            )
-            work_conformance = _work_profile_conformance(
-                inspection, "qualify" if action == "qualify" else "installed-runtime"
-            )
-            if work_conformance is not None:
-                request["work_conformance"] = work_conformance
-    plan = storage_service.profile_lifecycle(runtime_dir, "plan", request=request)
-    return {
-        "schema": "kungfu.profile-agent-plan/v1",
-        "corePlan": plan,
-        "decisionCard": _lifecycle_decision_card(action, plan),
-    }
-
-
-def lifecycle_apply(
-    runtime_dir: str | Path, core_plan: Mapping[str, Any], authorization_id: str
-) -> dict[str, Any]:
-    return storage_service.profile_lifecycle(
-        runtime_dir, "apply", plan=dict(core_plan), authorization_id=authorization_id
-    )
-
-
-def answer_decision(
-    card: Mapping[str, Any], choice: str, authorized_by: str
-) -> dict[str, Any]:
-    _validate_decision_card(card)
-    if choice not in card.get("choices", []):
-        raise ProfileSdkError(
-            "decision-choice-invalid",
-            "answer is not one of the decision card choices",
-            choices=card.get("choices", []),
-        )
-    actor = authorized_by.strip()
-    if not actor:
-        raise ProfileSdkError(
-            "decision-actor-required", "authorized_by must not be empty"
-        )
-    identity = {
-        "cardId": card["cardId"],
-        "choice": choice,
-        "authorizedBy": actor,
-        "requiredAuthority": card["requiredAuthority"],
-        "basis": card["basis"],
-    }
-    answer = {
-        "schema": DECISION_ANSWER_SCHEMA,
-        "authorizationId": _root(identity),
-        **identity,
-        "authorityVerification": "external-policy-required",
-    }
-    _validate_decision_answer(answer, card)
-    return answer
-
-
-def authorized_lifecycle_apply(
-    runtime_dir: str | Path,
-    agent_plan: Mapping[str, Any],
-    answer: Mapping[str, Any],
-) -> dict[str, Any]:
-    if agent_plan.get("schema") != "kungfu.profile-agent-plan/v1":
-        raise ProfileSdkError(
-            "agent-plan-invalid", "apply requires a Profile Agent plan"
-        )
-    card = agent_plan.get("decisionCard") or {}
-    core_plan = agent_plan.get("corePlan") or {}
-    expected_card = _lifecycle_decision_card(
-        str((core_plan.get("request") or {}).get("action") or ""), core_plan
-    )
-    if card.get("cardId") != expected_card.get("cardId"):
-        raise ProfileSdkError(
-            "decision-card-mismatch", "Agent plan decision card was altered"
-        )
-    _validate_decision_answer(answer, card)
-    if answer.get("choice") != "approve":
-        raise ProfileSdkError(
-            "decision-denied", "the Profile lifecycle plan was not approved"
-        )
-    if (answer.get("basis") or {}).get("planId") != core_plan.get("plan_id"):
-        raise ProfileSdkError(
-            "decision-basis-mismatch", "decision answer does not bind this Core plan"
-        )
-    return lifecycle_apply(
-        runtime_dir, core_plan, str(answer.get("authorizationId") or "")
-    )
-
-
-def authorize_current_lifecycle(
-    runtime_dir: str | Path,
-    action: str,
-    source: str | Path,
-    expected_plan_id: str,
-    choice: str,
-    authorized_by: str,
-) -> dict[str, Any]:
-    """Re-plan an exact source cut, then answer and apply its installed card."""
-
-    plan = lifecycle_plan(runtime_dir, action, source)
-    actual = str((plan.get("corePlan") or {}).get("plan_id") or "")
-    if not expected_plan_id or actual != expected_plan_id:
-        raise ProfileSdkError(
-            "lifecycle-plan-stale",
-            "Profile lifecycle plan changed; review a new decision card",
-            expectedPlanId=expected_plan_id,
-            actualPlanId=actual,
-        )
-    answer = answer_decision(plan["decisionCard"], choice, authorized_by)
-    return authorized_lifecycle_apply(runtime_dir, plan, answer)
-
-
-def semantic_diff(left: str | Path, right: str | Path) -> dict[str, Any]:
-    a = resolve_source(left)
-    b = resolve_source(right)
-    pa, pb = a["profile"], b["profile"]
-    categories = {
-        "display": _changes(pa, pb, ["title", "views"]),
-        "content": _changes(pa, pb, ["kfd1", "members"]),
-        "permission": _changes(pa, pb, ["permissions"]),
-        "authority": _changes(pa, pb, ["actions", "kfd3"]),
-        "evidence": _changes(pa, pb, ["kfd2", "qualification"]),
-        "migration": _changes(pa, pb, ["migrations"]),
-    }
-    cards = []
-    for category, authority in {
-        "permission": "workspace-profile-operator",
-        "authority": "profile-authority-owner",
-        "evidence": "evidence-policy-owner",
-        "migration": "workspace-data-owner",
-    }.items():
-        if categories[category]:
-            cards.append(
-                decision_card(
-                    f"profile-{category}-change",
-                    f"Approve or reject the Profile {category} change.",
-                    choices=["approve", "reject", "revise"],
-                    basis={"category": category, "changes": categories[category]},
-                    required_authority=authority,
-                    resume_command="rerun kungfu profile diff, then create a fresh lifecycle plan",
-                )
-            )
-    return {
-        "schema": "kungfu.profile-semantic-diff/v1",
-        "leftRoot": validate_source(left, "")["inspection"]["profile_suite_root"],
-        "rightRoot": validate_source(right, "")["inspection"]["profile_suite_root"],
-        "categories": categories,
-        "changedCategories": [key for key, rows in categories.items() if rows],
-        "decisionCards": cards,
-    }
-
-
-def action_catalog(source: str | Path, runtime_dir: str | Path) -> dict[str, Any]:
-    validated = validate_source(source, runtime_dir)
-    inspection = validated["inspection"]
-    registry = _read_ref_json(inspection, inspection["profile"]["actions"]["registry"])
-    _validate_action_registry(registry, inspection["profile"])
-    return {
-        "schema": "kungfu.profile-action-catalog/v1",
-        "profileId": inspection["profile"]["id"],
-        "profileSuiteRoot": inspection["profile_suite_root"],
-        "source": str(Path(source).resolve()),
-        "actions": registry["actions"],
-    }
-
-
-def _action_runtime_plan(
-    runtime_dir: str | Path, action: Mapping[str, Any]
-) -> tuple[dict[str, Any], dict[str, Any] | None]:
-    operation_id = str(action.get("runtimeOperation") or "")
-    try:
-        operation = runtime_broker.operation_definition(operation_id)
-    except (KeyError, ValueError) as error:
-        raise ProfileSdkError(
-            "action-runtime-operation-invalid", str(error), operationId=operation_id
-        ) from error
-    evidence = None
-    minimum_cut = None
-    if operation["operationClass"] != "storage-only":
-        evidence_path = runtime_broker.native_readiness_evidence_path(runtime_dir)
-        try:
-            evidence = runtime_broker.discover_native_readiness_evidence(runtime_dir)
-        except ValueError as error:
-            raise ProfileSdkError(
-                "runtime-evidence-invalid",
-                str(error),
-                evidencePath=str(evidence_path),
-            ) from error
-        if evidence is None:
-            raise ProfileSdkError(
-                "runtime-evidence-unavailable",
-                "live Profile action requires native durability evidence coordinates",
-                evidencePath=str(evidence_path),
-                operationId=operation_id,
-            )
-        minimum_cut = evidence["minimumCut"]
-        if (
-            "runtime.live-projection" in operation["requiredCapabilities"]
-            and evidence.get("projection") is None
-        ):
-            raise ProfileSdkError(
-                "runtime-evidence-incomplete",
-                "live projection action requires projection evidence coordinates",
-                evidencePath=str(evidence_path),
-                operationId=operation_id,
-            )
-    runtime_plan = runtime_broker.plan_operation(
-        operation_id,
-        workspace=runtime_broker.workspace_id(runtime_dir),
-        request_source="kfx",
-        minimum_cut=minimum_cut,
-    )
-    return runtime_plan, evidence
-
-
-def plan_action(
-    source: str | Path, runtime_dir: str | Path, action_id: str, input_value: Any
-) -> dict[str, Any]:
-    catalog = action_catalog(source, runtime_dir)
-    collaboration_value = collaboration(source, runtime_dir)
-    action = next((row for row in catalog["actions"] if row["id"] == action_id), None)
-    if action is None:
-        raise ProfileSdkError(
-            "action-not-found", f"Profile action not found: {action_id}"
-        )
-    state = storage_service.profile_lifecycle(
-        runtime_dir, "get", profile_id=catalog["profileId"]
-    )
-    if (
-        not state.get("activated")
-        or state.get("profile_suite_root") != catalog["profileSuiteRoot"]
-    ):
-        raise ProfileSdkError(
-            "profile-not-active", "Action requires this exact Profile root to be active"
-        )
-    identity = {
-        "profileId": catalog["profileId"],
-        "profileSuiteRoot": catalog["profileSuiteRoot"],
-        "action": action,
-        "input": input_value,
-        "stateRevision": state["revision"],
-    }
-    if collaboration_value["declared"]:
-        intent = next(
-            (
-                row
-                for row in collaboration_value["intents"]
-                if row["actionId"] == action_id
-            ),
-            None,
-        )
-        if intent is None:
-            raise ProfileSdkError(
-                "collaboration-action-closure",
-                "declared collaboration has no intent for this action",
-            )
-        identity.update(
-            {
-                "collaborationRoot": collaboration_value["collaborationRoot"],
-                "closureRoot": collaboration_value["closureRoot"],
-                "intentId": intent["id"],
-            }
-        )
-    missing_capabilities = sorted(
-        set(action["requiredCapabilities"]) - set(state.get("granted_permissions", []))
-    )
-    if missing_capabilities:
-        raise ProfileSdkError(
-            "action-capability-not-granted",
-            "Action requires Profile capabilities that are not active",
-            missingCapabilities=missing_capabilities,
-        )
-    runtime_plan, runtime_evidence = _action_runtime_plan(runtime_dir, action)
-    identity["runtimeOperation"] = runtime_plan["operation"]["id"]
-    if runtime_evidence is not None:
-        identity.update(
-            {
-                "runtimePlan": runtime_plan,
-                "runtimeEvidence": runtime_evidence,
-            }
-        )
-    plan = {
-        "schema": ACTION_PLAN_SCHEMA,
-        "planId": _root(identity),
-        **identity,
-        "source": catalog["source"],
-        "runtimePlan": runtime_plan,
-        "runtimeEvidence": runtime_evidence,
-        "requiresAuthorization": action["authorityClass"] != "none",
-        "effects": action.get("effects", []),
-    }
-    if plan["requiresAuthorization"]:
-        plan["decisionCard"] = decision_card(
-            "profile-action-authorization",
-            f"Authorize Profile action {action_id} for the exact active root.",
-            choices=["approve", "deny"],
-            basis={
-                "planId": plan["planId"],
-                "profileSuiteRoot": catalog["profileSuiteRoot"],
-            },
-            required_authority=action["authorityClass"],
-            resume_command="answer this card, then invoke with the exact action plan and decision answer",
-        )
-    return plan
-
-
-def invoke_action(
-    runtime_dir: str | Path, plan: Mapping[str, Any], authorization_id: str | None
-) -> dict[str, Any]:
-    if plan.get("schema") != ACTION_PLAN_SCHEMA:
-        raise ProfileSdkError(
-            "action-plan-invalid", "invoke requires a Profile action plan"
-        )
-    refreshed = plan_action(
-        str(plan.get("source") or ""),
-        runtime_dir,
-        str((plan.get("action") or {}).get("id") or ""),
-        plan.get("input"),
-    )
-    if refreshed.get("planId") != plan.get("planId"):
-        raise ProfileSdkError(
-            "action-plan-stale", "Profile source or active state changed after planning"
-        )
-    if refreshed.get("runtimePlan") != plan.get("runtimePlan") or refreshed.get(
-        "runtimeEvidence"
-    ) != plan.get("runtimeEvidence"):
-        raise ProfileSdkError(
-            "action-plan-stale", "runtime execution material changed after planning"
-        )
-    action = plan.get("action") or {}
-    if plan.get("requiresAuthorization") and not authorization_id:
-        raise ProfileSdkError(
-            "authorization-required", "Profile action requires an authorization id"
-        )
-    state = storage_service.profile_lifecycle(
-        runtime_dir, "get", profile_id=plan["profileId"]
-    )
-    if state.get("profile_suite_root") != plan.get("profileSuiteRoot") or state.get(
-        "revision"
-    ) != plan.get("stateRevision"):
-        raise ProfileSdkError(
-            "action-plan-stale", "active Profile state changed after action planning"
-        )
-    if action.get("runner") == "kfx-member":
-
-        def invoke_kfx(_activation: Mapping[str, Any]) -> dict[str, Any]:
-            receipt = invoke_member_adapter(
-                str(plan.get("source") or ""),
-                runtime_dir,
-                str(action.get("operation") or ""),
-                str(action.get("id") or ""),
-                plan.get("input"),
-                authorized_action=True,
-            )
-            result = receipt.get("result")
-            if not isinstance(result, Mapping):
-                raise ProfileSdkError(
-                    "member-adapter-result-invalid",
-                    "Profile action member must return an object result",
-                )
-            return {
-                **result,
-                "memberReceipt": {
-                    key: value for key, value in receipt.items() if key != "result"
-                },
-            }
-
-        callback = invoke_kfx
-    elif action.get("runner") == "profile-lifecycle":
-        lifecycle_action = action.get("operation")
-        if lifecycle_action not in {"qualify", "activate", "remove"}:
-            raise ProfileSdkError(
-                "action-operation-unsupported",
-                "unsupported lifecycle action operation",
-            )
-        source_value = plan.get("source")
-        lifecycle_source = str(source_value) if source_value is not None else None
-        lifecycle_values: dict[str, Any] = {}
-        if lifecycle_action == "remove":
-            lifecycle_values["profile_id"] = plan["profileId"]
-            lifecycle_source = None
-        core = lifecycle_plan(
-            runtime_dir, lifecycle_action, lifecycle_source, **lifecycle_values
-        )["corePlan"]
-
-        def invoke_lifecycle(_activation: Mapping[str, Any]) -> dict[str, Any]:
-            return {
-                "coreReceipt": lifecycle_apply(
-                    runtime_dir, core, authorization_id or "action-policy:none"
-                )
-            }
-
-        callback = invoke_lifecycle
-    else:
-        raise ProfileSdkError(
-            "action-runner-unsupported", "unsupported Profile action runner"
-        )
-
-    evidence = plan.get("runtimeEvidence")
-    readiness_authority = None
-    runtime_home = str(Path(runtime_dir).expanduser().resolve().parent)
-    if isinstance(evidence, Mapping):
-        try:
-            readiness_authority = runtime_broker.native_readiness_authority(evidence)
-        except (TypeError, ValueError) as error:
-            raise ProfileSdkError("runtime-evidence-invalid", str(error)) from error
-        runtime_home = str(evidence["runtimeHome"])
-    broker = runtime_broker.RuntimeCapabilityBroker.for_process(
-        runtime_home,
-        str(runtime_dir),
-        readiness_authority=readiness_authority,
-    )
-    runtime_plan = plan.get("runtimePlan")
-    if not isinstance(runtime_plan, Mapping):
-        raise ProfileSdkError(
-            "action-runtime-plan-invalid", "Profile action has no runtime plan"
-        )
-    try:
-        runtime_receipt = broker.invoke(runtime_plan, callback)
-    except ValueError as error:
-        raise ProfileSdkError("action-runtime-plan-invalid", str(error)) from error
-    result = runtime_receipt.get("result")
-    if not runtime_receipt.get("accepted") or not isinstance(result, Mapping):
-        return {
-            "schema": ACTION_RECEIPT_SCHEMA,
-            "planId": plan["planId"],
-            "authorizationId": authorization_id,
-            "runtimeReceipt": runtime_receipt,
-            "coreReceipt": None,
-            "verified": False,
-        }
-    return {
-        "schema": ACTION_RECEIPT_SCHEMA,
-        "planId": plan["planId"],
-        "authorizationId": authorization_id,
-        "runtimeReceipt": runtime_receipt,
-        **result,
-        "verified": True,
-    }
-
-
-def authorized_action_invoke(
-    runtime_dir: str | Path,
-    plan: Mapping[str, Any],
-    answer: Mapping[str, Any] | None,
-) -> dict[str, Any]:
-    if plan.get("schema") != ACTION_PLAN_SCHEMA:
-        raise ProfileSdkError(
-            "action-plan-invalid", "invoke requires a Profile action plan"
-        )
-    if not plan.get("requiresAuthorization"):
-        return invoke_action(runtime_dir, plan, None)
-    refreshed = plan_action(
-        str(plan.get("source") or ""),
-        runtime_dir,
-        str((plan.get("action") or {}).get("id") or ""),
-        plan.get("input"),
-    )
-    card = plan.get("decisionCard") or {}
-    if card.get("cardId") != (refreshed.get("decisionCard") or {}).get("cardId"):
-        raise ProfileSdkError(
-            "decision-card-mismatch", "action plan decision card was altered"
-        )
-    if not answer:
-        raise ProfileSdkError(
-            "decision-answer-invalid", "action invoke requires a decision answer"
-        )
-    _validate_decision_answer(answer, card)
-    if answer.get("choice") != "approve":
-        raise ProfileSdkError("decision-denied", "the Profile action was not approved")
-    if (answer.get("basis") or {}).get("planId") != plan.get("planId"):
-        raise ProfileSdkError(
-            "decision-basis-mismatch", "decision answer does not bind this action plan"
-        )
-    return invoke_action(runtime_dir, plan, str(answer.get("authorizationId") or ""))
-
-
-def _source_plan_identity(
-    brief: Any, destination: str, files: Mapping[str, bytes]
-) -> dict[str, Any]:
-    return {
-        "brief": brief,
-        "destination": destination,
-        "files": [
-            {"path": path, "sha256": _sha256(data)}
-            for path, data in sorted(files.items())
-        ],
-    }
-
-
-def _lifecycle_decision_card(action: str, plan: Mapping[str, Any]) -> dict[str, Any]:
-    return decision_card(
-        "profile-lifecycle-authorization",
-        f"Authorize the exact {action} plan for this Profile root.",
-        choices=["approve", "deny"],
-        basis={
-            "planId": plan.get("plan_id"),
-            "basis": plan.get("basis"),
-            "effects": plan.get("effects"),
-        },
-        required_authority="workspace-profile-operator",
-        resume_command="kungfu profile decide <plan.json> --choice approve --authorized-by <actor> --out <answer.json> --json; kungfu profile apply <plan.json> --authorization-file <answer.json> --json",
-    )
-
-
-def _validate_decision_card(card: Mapping[str, Any]) -> None:
-    _validate_sdk_value("decisionCardSchema", dict(card), "decision card")
-    if card.get("schema") != DECISION_CARD_SCHEMA or card.get("status") != "open":
-        raise ProfileSdkError(
-            "decision-card-invalid", "answer requires an open decision card"
-        )
-    identity = {
-        "kind": card.get("kind"),
-        "question": card.get("question"),
-        "choices": card.get("choices"),
-        "basis": card.get("basis"),
-        "requiredAuthority": card.get("requiredAuthority"),
-    }
-    if card.get("cardId") != _root(identity):
-        raise ProfileSdkError(
-            "decision-card-tampered", "decision card no longer matches its identity"
-        )
-
-
-def _validate_decision_answer(
-    answer: Mapping[str, Any], card: Mapping[str, Any]
-) -> None:
-    _validate_decision_card(card)
-    _validate_sdk_value("decisionAnswerSchema", dict(answer), "decision answer")
-    identity = {
-        "cardId": answer.get("cardId"),
-        "choice": answer.get("choice"),
-        "authorizedBy": answer.get("authorizedBy"),
-        "requiredAuthority": answer.get("requiredAuthority"),
-        "basis": answer.get("basis"),
-    }
-    if answer.get("cardId") != card.get("cardId"):
-        raise ProfileSdkError(
-            "decision-answer-mismatch", "decision answer targets another card"
-        )
-    if answer.get("requiredAuthority") != card.get("requiredAuthority"):
-        raise ProfileSdkError(
-            "decision-authority-mismatch", "decision answer changes required authority"
-        )
-    if answer.get("choice") not in card.get("choices", []):
-        raise ProfileSdkError(
-            "decision-choice-invalid", "decision answer choice is no longer offered"
-        )
-    if answer.get("basis") != card.get("basis"):
-        raise ProfileSdkError(
-            "decision-basis-mismatch", "decision answer changes the card basis"
-        )
-    if not str(answer.get("authorizedBy") or "").strip():
-        raise ProfileSdkError("decision-actor-required", "decision answer has no actor")
-    if answer.get("authorizationId") != _root(identity):
-        raise ProfileSdkError(
-            "decision-answer-tampered", "decision answer no longer matches its identity"
-        )
-
-
-def _normalize_brief(
-    brief: Mapping[str, Any],
-) -> tuple[dict[str, Any], list[dict[str, Any]]]:
-    allowed = {
-        "schema",
-        "id",
-        "title",
-        "version",
-        "purposes",
-        "permissions",
-        "identity",
-        "evidence",
-        "migration",
-        "collaboration",
-    }
-    if set(brief) - allowed:
-        raise ProfileSdkError(
-            "brief-field-unknown",
-            "brief contains fields not owned by the installed contract",
-            fields=sorted(set(brief) - allowed),
-        )
-    if brief.get("schema") != BRIEF_SCHEMA:
-        raise ProfileSdkError("brief-schema-invalid", f"brief must use {BRIEF_SCHEMA}")
-    profile_id = str(brief.get("id") or "").strip()
-    title = str(brief.get("title") or "").strip()
-    version = str(brief.get("version") or "1.0.0").strip()
-    if not profile_id or not _TOKEN.fullmatch(profile_id):
-        raise ProfileSdkError(
-            "profile-id-invalid", "brief.id must be a safe Profile token"
-        )
-    if not title:
-        raise ProfileSdkError("profile-title-required", "brief.title must not be empty")
-    cards = []
-    checks = [
-        (
-            "identity",
-            "authority",
-            "identity-authority",
-            ["workspace-owner", "declared-source-owner"],
-        ),
-        (
-            "evidence",
-            "strength",
-            "evidence-strength",
-            ["reported-with-references", "independently-observed"],
-        ),
-        (
-            "migration",
-            "mode",
-            "migration-mode",
-            ["additive", "explicit-destructive-plan"],
-        ),
-    ]
-    for section, field, kind, choices in checks:
-        value = brief.get(section)
-        selected = value.get(field) if isinstance(value, Mapping) else None
-        if selected not in choices:
-            cards.append(
-                decision_card(
-                    kind,
-                    f"Choose the Profile {section} {field} boundary.",
-                    choices=choices,
-                    basis={
-                        "profileId": profile_id,
-                        "briefSha256": _sha256(_canonical(brief)),
-                        "supplied": selected,
-                    },
-                    required_authority="profile-author",
-                    resume_command="update brief.json, then rerun kungfu profile scaffold",
-                )
-            )
-    migration = brief.get("migration")
-    if (
-        isinstance(migration, Mapping)
-        and migration.get("mode") == "explicit-destructive-plan"
-    ):
-        cards.append(
-            decision_card(
-                "destructive-migration",
-                "A destructive migration requires a separate bounded migration plan.",
-                choices=["switch-to-additive", "prepare-separate-migration-plan"],
-                basis={"profileId": profile_id, "migration": dict(migration)},
-                required_authority="workspace-data-owner",
-                resume_command="revise the brief, then rerun kungfu profile scaffold",
-            )
-        )
-    normalized = {
-        "schema": BRIEF_SCHEMA,
-        "id": profile_id,
-        "title": title,
-        "version": version,
-        "purposes": sorted(
-            set(str(v) for v in brief.get("purposes", ["operator-review"]))
-        ),
-        "permissions": sorted(set(str(v) for v in brief.get("permissions", []))),
-        "identity": dict(brief.get("identity") or {}),
-        "evidence": dict(brief.get("evidence") or {}),
-        "migration": dict(brief.get("migration") or {}),
-    }
-    collaboration = brief.get("collaboration")
-    if collaboration is not None:
-        if not isinstance(collaboration, Mapping):
-            raise ProfileSdkError(
-                "collaboration-brief-invalid",
-                "brief.collaboration must be an object when KFD-3 is requested",
-            )
-        artifact = {
-            "schema": COLLABORATION_SCHEMA,
-            "profileId": profile_id,
-            "value": {
-                "summary": collaboration.get("summary"),
-                "participantBenefits": collaboration.get("participantBenefits", []),
-            },
-            "participants": collaboration.get("participants", []),
-            "intents": [],
-            "constraints": collaboration.get("constraints", []),
-            "knownLimits": collaboration.get("knownLimits", []),
-            "presentation": {"mode": "generic", "homeViewId": None},
-        }
-        _validate_sdk_value(
-            "collaborationSchema", artifact, "Profile brief collaboration"
-        )
-        normalized["collaboration"] = artifact
-    if not cards:
-        _validate_sdk_value("briefSchema", normalized, "Profile brief")
-    return normalized, cards
+from kungfu._profile_sdk.lifecycle import (  # noqa: E402
+    export_source_bundle as export_source_bundle,
+    source_import_plan as source_import_plan,
+    authorized_source_import as authorized_source_import,
+    lifecycle_plan as lifecycle_plan,
+    lifecycle_apply as lifecycle_apply,
+    answer_decision as answer_decision,
+    authorized_lifecycle_apply as authorized_lifecycle_apply,
+    authorize_current_lifecycle as authorize_current_lifecycle,
+)
+from kungfu._profile_sdk.action import (  # noqa: E402
+    semantic_diff as semantic_diff,
+    action_catalog as action_catalog,
+    _action_runtime_plan as _action_runtime_plan,
+    plan_action as plan_action,
+    invoke_action as invoke_action,
+    authorized_action_invoke as authorized_action_invoke,
+    _source_plan_identity as _source_plan_identity,
+    _lifecycle_decision_card as _lifecycle_decision_card,
+    _validate_decision_card as _validate_decision_card,
+    _validate_decision_answer as _validate_decision_answer,
+    _normalize_brief as _normalize_brief,
+)
+from kungfu._profile_sdk.intent import (  # noqa: E402
+    intent_inspect as intent_inspect,
+    intent_advise as intent_advise,
+    intent_plan as intent_plan,
+    intent_apply as intent_apply,
+    intent_verify as intent_verify,
+    authorize_current_intent as authorize_current_intent,
+)
+from kungfu._profile_sdk.kfd3 import (  # noqa: E402
+    _kfd3_operations as _kfd3_operations,
+    _earn_kfd3 as _earn_kfd3,
+    build_kfd3_release_manifest as build_kfd3_release_manifest,
+    kfd3_status as kfd3_status,
+    kfd3_qualification_plan as kfd3_qualification_plan,
+    qualify_kfd3 as qualify_kfd3,
+    authorize_kfd3_qualification as authorize_kfd3_qualification,
+    verify_kfd3 as verify_kfd3,
+    qualify_source as qualify_source,
+)
+
+for _public_callable in (
+    export_source_bundle,
+    source_import_plan,
+    authorized_source_import,
+    lifecycle_plan,
+    lifecycle_apply,
+    answer_decision,
+    authorized_lifecycle_apply,
+    authorize_current_lifecycle,
+    semantic_diff,
+    action_catalog,
+    plan_action,
+    invoke_action,
+    authorized_action_invoke,
+    intent_inspect,
+    intent_advise,
+    intent_plan,
+    intent_apply,
+    intent_verify,
+    authorize_current_intent,
+    build_kfd3_release_manifest,
+    kfd3_status,
+    kfd3_qualification_plan,
+    qualify_kfd3,
+    authorize_kfd3_qualification,
+    verify_kfd3,
+    qualify_source,
+):
+    _public_callable.__module__ = __name__
